cytnx_bench: merge algorithm scripts into their test files
Each script's run_one(chi, L) now lives directly in its test_<name>.py
file instead of a separate, non-test module imported via a
sys.path.insert(0, os.path.join(os.path.dirname(__file__), "..")) hack.
test_dmrg.py (which combined the dense and symmetric DMRG cases) is
replaced by test_dmrg_dense.py and test_dmrg_symmetric.py, one file
per algorithm, matching the layout of every other script in this
directory.

Every benchmark test now asserts the returned energy against a
REFERENCE_ENERGIES dict covering the full (chi, L) grid (sourced from
a full-grid run of each script), replacing the previous single
fixed-point regression check plus an isfinite-only full-grid sweep.
The grid is parametrized with two stacked
@pytest.mark.parametrize("chi", ...) / @pytest.mark.parametrize("length", ...)
decorators instead of a single parametrize over a combined (chi, length)
list, and limit_memory is applied as a plain decorator on a dedicated,
parameter-free *_memory test pinned to the (chi=16, L=20) point instead
of being attached via pytest.param(..., marks=...).
